### 2020-10-16 #2757 @whedon
[REVIEW]: The Pulsar Signal Simulator: A Python package for simulating radio signal data from pulsars
**Submitting author:** @Hazboun6 (<a href="http://orcid.org/0000-0003-2742-3321">Jeffrey Hazboun</a>)
**Repository:** <a href="https://github.com/PsrSigSim/PsrSigSim" target ="_blank">https://github.com/PsrSigSim/PsrSigSim</a>
**Version:** v1.0.1
