### 2020-10-16 #2757 @mbobra
👋 @JBorrow @arjunsavel Thank you so much for agreeing to review this submission! You'll see your reviewer checklists above. Please feel free to ask questions here on this review thread.

Please note that this article is in collaboration with AAS publishing (the submitting authors submitted a companion paper to *The Astrophysical Journal*). As part of this collaboration, AAS publishing makes a small donation to the running costs of JOSS. Please take a look at the blog posts [about the AAS-JOSS collaboration](https://blog.joss.theoj.org/2018/12/a-new-collaboration-with-aas-publishing) and [JOSS' financial model](https://joss.theoj.org/about#costs). Please let me know if you have any issues with this.
