### 2020-10-17 PsrSigSim/PsrSigSim#156 @JBorrow
Module missing from install_requires and conflicts with `pint` library.
In a clean install on python3.8.0, from pip, I get the following error:

```
Traceback (most recent call last):
  File "getting_started.py", line 4, in <module>
    import psrsigsim as pss
  File "/private/tmp/env/lib/python3.8/site-packages/psrsigsim/__init__.py", line 15, in <module>
    from . import io
  File "/private/tmp/env/lib/python3.8/site-packages/psrsigsim/io/__init__.py", line 4, in <module>
    from .psrfits import PSRFITS
  File "/private/tmp/env/lib/python3.8/site-packages/psrsigsim/io/psrfits.py", line 7, in <module>
    from packaging import version
ModuleNotFoundError: No module named 'packaging'
```

This is of course resolved by installing `packaging`, but this should really be part of the `install_requires`.

One of your dependencies, `pint-pulsar`, has a concerning name conflict with the very popular `pint` symbolic unit library. This caused very confusing problems for me when running the tests. If this package is not really required (although based on its name, I guess it is), then it would be good to remove it. If not, you should definitely point this problem out clearly in https://psrsigsim.readthedocs.io/en/latest/contributing.html.

This issue is part of my review for JOSS.
