### 2020-10-17 PsrSigSim/PsrSigSim#157 @JBorrow
Test failures and large amount of warnings on python3.8.0
I'm getting a few test failures and a lot of warnings on my install of python3.8.0, on MacOS. I have attached the output - can you please tell me what I'm doing wrong?

This is in a fresh virtual environment, and I did:

```
pip install psrsigsim
pip install -r requirements.txt
pip install -r requirements_dev.txt
```

before

```
python3 -m pytest tests
```

The failures are mainly in test_io.py.

I would also like to see the warnings in the tests, particularly the DeprecationWarnings, addressed.

[test_output.txt](https://github.com/PsrSigSim/PsrSigSim/files/5396446/test_output.txt)

This issue is part of my JOSS review.
