### 2020-10-17 PsrSigSim/PsrSigSim#158 @JBorrow
AstropyDeprecationWarning
When running the code, as in the examples, I continually receive the following warning:

`WARNING: AstropyDeprecationWarning: The truth value of a Quantity is ambiguous. In the future this will raise a ValueError. [astropy.units.quantity]`

It would be good to track down the origin of this, and fix it. I suppose what's happening here is that an astropy quantity is being checked to see if it is non-zero; in that case this should be easily resolved by checking the value against zero.

I am running `astropy-3.2.3`

This issue is part of my JOSS review.
