### 2020-10-17 PsrSigSim/PsrSigSim#159 @JBorrow
Documentation Issues
I've found a few small issues with the docs:

Usage page: https://psrsigsim.readthedocs.io/en/latest/usage.html. This page is very short and probably unecessary. It also includes a TBD link: "An advanced example, based on [link to Brent’s paper] can be found on MyBinder:"

General issue: it would be good if, during the narrative docs, you could link to the relevant pieces in the API docs, e.g. using the `:mod:`, `:func:`, and `:obj:` syntax.

First example
-------------

In the first example: https://psrsigsim.readthedocs.io/en/latest/tutorial_1.html#the-pulsar, when I used the first set of the code, I found that the code started downloading files. I wasn't warned about this in the tutorial, and it didn't tell me where the files would go. Seeing as you are printing about downloading the files anyway, it would be good to add a description and tell us where they are so we can delete them later should we wish.

Typo: "Out pulsar needs a period (s)" should be "Our pulsar needs a period (s)"

In your example code, the formatting is not consistent - not critical, but it would be nice to have this code formatted the same (e.g. spacing around keyword argument names or not).

Third Tutorial
--------------

Typo in the definition of the gaussian profiles, `mulit_gauss_prof` should be `multi_gauss_prof`.

The organisation of your example data is not consistent. At the top level of your repository you have a `data` folder, but then within the package itself you have another `data` folder. As the data used in this example is 'example' data, maybe it is best to move it to the example folder?

Typo: `Portrat` should be `Portrait` under Data Portraits. This occurs later, too.

Another typo, `mulit-gaussian`, instead of `multi-gaussian`, under 'Simulations'.

Fourth Tutorial
--------------

It would be nice to have the links to the papers (Zhu et al. 2015, etc.) rather than just these short citations.

Typo: `frequeny^-2` in Pulse Dispersion.

This issue is part of my JOSS review.
