### 2020-10-17 PsrSigSim/PsrSigSim#160 @JBorrow
Paper Review
I only have two short comments on the JOSS paper that you have submitted:

1. The citation `[Shapiro-Albert, 2020]` should not have those square brackets - I think you mean to use `@PaperYear` (i.e. not `[@PaperYear]` as used for the 'regular' bracketed citations).

2. I would like to see a short review of the state of the field, with a focus on answering the following two questions:
  a. Is `PsrSigSim` the first software package of its kind? If not, list other competing packages.
  b. What is particularly novel about `PsrSigSim`? This could be anything from the inclusion of new functionality relative to other packages, or just the fact that it has a nice object-oriented interface.

Other than this, the paper looks good to me, especially considering this will be combined with an ApJ publication.

This issue is part of my JOSS review.
