### 2020-10-17 #2757 @JBorrow
The authors have presented a new package for the simulation and modelling of pulsar observations. It has a clean object-oriented interface, and exceptionally detailed narrative documentation and tutorial content. I found the interface particularly intuitive, and it was great (as somebody who is a 'pure' simulator) to play around with the different telescopes. I can see clearly both the scientific as well as educational benefit of such a package. As somebody not directly within the field of radio astronomy, I cannot comment on the scientific merit of the package or its results, but I will leave this to those reviewing the ApJ paper.

I have opened the following issues within the main repository:

+ https://github.com/PsrSigSim/PsrSigSim/issues/156
+ https://github.com/PsrSigSim/PsrSigSim/issues/157
+ https://github.com/PsrSigSim/PsrSigSim/issues/158
+ https://github.com/PsrSigSim/PsrSigSim/issues/159
+ https://github.com/PsrSigSim/PsrSigSim/issues/160

with various concerns. Most of these are minor and should be a relatively quick fix (e.g. typos, adding a section to the paper looking at other software packages in the field). My one major concern, though, is the large amount of warnings produced when running the examples and tests. In particular, when running the test suite, I saw over 1000 warnings of various types (Runtime warnings from taking negative square roots and deprecation warnings). I would like to have these resolved before being confident in accepting this package, particularly as they are so numerous.
