### 2020-10-17 PsrSigSim/VersionZeroPointZero#157 @JBorrow
It appears that the test failures are connected to astropy `4.0.0` somehow ending up installed on my machine, despite me specifying 3.3.2. However, the tests still generate a large number of warnings.
