### 2020-10-19 PsrSigSim/PsrSigSim#161 @arjunsavel
Paper Review
The submitted JOSS paper for this project looks great! Most of my suggestions are nitpicky and grammar-related:

- [x] Nobel prize —> Nobel Prize
- [x] these neutron stars —> neutron stars
- [x] gravitational wave observatory —> gravitational wave observatory,
- [x] The PsrSigSim —>  PsrSigSim
    - Unless you feel strongly otherwise about naming/reference conventions.
- [x] “And the propagation effects upon that signal from the interstellar medium —> “and the effects of the interstellar material on the signal propagation”
- [x] Unhyphenate the “-objects”
- [x] “signal processing techniques” —> “signal-processing techniques”
- [x] “frequency dependent effects” —> “frequency-dependent effects”
- [x] I'd add a citation for the [PSRFITS paper.](https://www.cambridge.org/core/journals/publications-of-the-astronomical-society-of-australia/article/psrchive-and-psrfits-an-open-approach-to-radio-pulsar-data-storage-and-analysis/1FACAD0DF4F55564BBF28041F33C8BB4)
- [x] “data processing pipelines” —> “data-processing pipelines”
- [x] “They also allow users” —> “It also allows users”

Additionally, I'll echo the sentiment in #160 re: reviewing the state of the field. I believe this would be most appropriate before the last sentence of the first page of the paper (i.e., before "The end-to-end simulation...").

This issue is part of my JOSS review.
