### 2020-10-19 PsrSigSim/PsrSigSim#162 @arjunsavel
Documentation Review
The documentation of this package is very comprehensive, and I was quickly able to intuit the code's core functionality. Of particular note was the "backyard telescope" example — it certainly highlighted the educational promise of this tool. I was able to reproduce the plots within the documentation without error. My comments are below; similarly to the paper, they are mostly related to grammar/consistency:

**Overall**
- [x] I'd suggest general consistency in italicizing the package name/using the package name in the code font, and writing "Github" as "GitHub"

**Contributing**
- [x] Link the GitHub repository

**Introductory tutorial 1**
- [x] Here and elsewhere, I'd make sure to add y-axis labels (even if they are arbitrary flux units)
- [x] Out pulsar —> our pulsar
- [x] Near the line introducing GBT: 9 —> (

**Fold Mode: Introductory Tutorial 2**
- [x] long term —> long-term
- [x] And then —> and then
- [x] Hyphenate "-style"

**Pulse profiles: Introductory tutorials**
- [x] Hyphenate two-dimensional
- [x] Add a ylabel (even if arbitrary!) to plots
- [x] I would consider a brief discussion of intensity units in general (or a pointer to a resource), given the potential for confusion.
- [x] maintianed —> maintained
- [x] lower frequency —> lower-frequency
- [x] high frequency —> high-frequency
- [x] "is a single Gauassian" —> "is a single Gaussian"
- [x] "multiple Gaussian defineing is profile" —> "multiple Gaussians defining its profile"
- [x] "resutls" —> "results"
- [x] "intial" —> "initial"
- [x] "mulit-gaussian" —> "multi-Gaussian"

**ISM delays**
- [x] I would mention that paths relative to root in the GitHub repository
- [x] 20 minute long —> 20-minute-long
- [x] The line including "(here 2048 samples across the pulse period" needs an end parens
- [x] intilialize —> initialize
- [x] frequeny —> frequency
- [x] "specifically frequency dependent delay" —> "specifically frequency-dependent delays"
- [x] "eponential" —> "exponential"
- [x] "dierectly shifting" —> "directly shifting"

**Telescopes**
- [x] 20 minute long —> 20-minute-long
- [x] "intilialize" —> "initialize"
- [x] "Here we will show to use" —> "Here we will show how to use"
- [x] "somebodies backyard" —> "somebody's backyard"
- [x] Add end parens for the  "(here 2048 samples across the pulse period"  line
- [x] A single plot comparing all the telescopes against each other might be nice!
- [x] aperature —> aperture

**Pulse nulling**
- [x] Period after "33%"
- [x] "exmample" —> "example"
- [x] exampe —> example
- [x] Period after pulsar
- [x] "lets" —> "let's"

This issue is part of my JOSS review.
