### 2020-10-19 PsrSigSim/PsrSigSim#164 @arjunsavel
Allow for user-input random seed
There are a few modules (`pulsar.pulsar`,` utils.utils`, `telescope.receiver`) where either the `numpy.random` module is explicitly called or samples are drawn from a normal distribution. In the interest of reproducibility (or exploring the effects of random noise on final results), I recommend allowing for user-input random seeds to initialize any random processes used. I would not consider implementation of this necessary for JOSS acceptance, but I would recommend a brief acknowledgment of plots in documentation not being exactly reproducible due to random processes.

This issue is part of my JOSS review.
