### 2020-10-19 #2757 @arjunsavel
I've  completed my review:

With PsrSigSim, the authors have provided a well-documented and well-organized package for simulating pulsar observations. I don't specialize in radio astronomy, but to the extent of my knowledge, I believe it to be of scientific merit.

Of the issues that I have opened in PsrSigSim repository, the below are relevant for acceptance. Many are related to grammar and consistency.
- https://github.com/PsrSigSim/PsrSigSim/issues/164
- https://github.com/PsrSigSim/PsrSigSim/issues/163
- https://github.com/PsrSigSim/PsrSigSim/issues/162
- https://github.com/PsrSigSim/PsrSigSim/issues/161

For https://github.com/PsrSigSim/PsrSigSim/issues/164, a full implementation of a user-input random seed would not be necessary for acceptance; however, I would recommend briefly addressing effects of randomness on the documentation plots.

I reproduced the errors (and fixes) on installation/testing that @JBorrow reported above (running Python 3.7.6).

I also wanted to touch base about the authorship list — I noticed that a few people are listed in the paper but not in AUTHORS.RST (Nathan Garver-Daniels, Maura A. McLaughlin).

Overall, the package looks great to me, and I'll be signed off pending the resolution of the above issues! Looking forward to the companion ApJ paper.
