### 2020-10-19 #2757 @Hazboun6
Thank you @JBorrow and @arjunsavel for your speedy reviews! We're looking over the Issues you both posted now and we'll let you know when we've been able to attend to them. Hopefully we'll be able to make most changes this week.
