### 2020-10-21 PsrSigSim/VersionZeroPointZero#158 @paulthebaker
At least one instance is in `telescope.observe()` ln. 93 when we check for subintegration
```
if signal.sublen:
```
