### 2020-10-21 PsrSigSim/VersionZeroPointZero#156 @paulthebaker
In anaconda python 3.8, all available `astropy` versions are 4.0+.  Astropy v3 should work with Python 3.8, but I can't check this in my usual workflow.  I can't find an official list of supported python versions for Astropy v3.2 after quick search.  Are we sure it includes v3.8?

We appear to be held at `astropy<4` by [nanograv/PINT#651](https://github.com/nanograv/PINT/issues/651).  Does this hold us at `python<3.8` too?

In `setup.py` we specify python 3.5, 3.6, and 3.7 as compatible versions and not python 3.8.  Do we need to add v3.8?
