### 2020-10-21 PsrSigSim/VersionZeroPointZero#156 @JBorrow
Python3.8 compatibility is not critical right now, but I would recommend you add it when you can. My major concern is the name of pint - installing this alongside the `pint` unit library causes everything to break. As this is a requirement of your code and people will probably have `pint` installed already, it would be good to either make a strong note of this in the docs, or to remove it as a dependency if you can get away with it.
