### 2020-10-21 PsrSigSim/VersionZeroPointZero#156 @bshapiroalbert
@paulthebaker you are correct that we were held to Astropy v3 by `PINT`, but I've just checked and it appears that they have updated their requirements for Astropy to: `astropy>=3.2,!=4.0.1,!=4.0.1.post1`, so I believe we could use the same requirements which would allow us to use python 3.8, but I should check (and @Hazboun6 may know) if this will cause issues with other packages currently required, particularly `pdat` and `fitsio`, or if it will cause issues with anything that currently relies on Astropy.

We do require `pint-pulsar` to handle a number of more complex calculations as other saving utilities related to the `io` class. I don't think we can get away with removing it as a dependency unfortunately, so we will make sure to make a strong note of this issue with the `pint` unit library in the docs.
