### 2020-10-21 PsrSigSim/VersionZeroPointZero#156 @Hazboun6
@bshapiroalbert I believe that the `!=` means "not equal to", so unfortunately I don't think they have fixed this. In fact I recently ran into the same issue that @JBorrow saw in Issue #157 (for a different project) where `astropy` 4.0 was installed automatically and then threw a number of errors.

[Edit] Sorry I misunderstood @bshapiroalbert . I didn't realize there were versions `>4.0.1.post1`.
