### 2020-10-21 PsrSigSim/VersionZeroPointZero#157 @Hazboun6
Moving some comments over here, as they seem more appropriate to this issue...

I think that allowing `Python 3.8` shouldn't be an issue (in theory). So I'll add it to the `Travis` configuration `Yaml`. At least then we'll know where issues might arise.
