### 2020-10-21 PsrSigSim/VersionZeroPointZero#160 @Hazboun6
@JBorrow This citation is a place holder for the AAS paper. In AAS-style guidelines they ask that you write in prep work directly into the paper, rather than add a separate citation. I suppose I should just ask the JOSS folks what the best option is for this.

A state of the field review is a great idea! There is at least one other package in `c`, but nothing in Python.
