### 2020-10-21 #2757 @Hazboun6
@mbobra One point that @JBorrow made in [Issue 160](https://github.com/PsrSigSim/PsrSigSim/issues/160) was about a reference to the AAS paper to which this work is linked. Is there a JOSS standard for how you'd like us to refer to that paper? I'm imagining it will probably take longer to appear/publish. Would you like it as `.bib` entry. Currently it is written as `[Shapiro-Albert,2020`, but I just realized that if I was followng the AAS style it should be `[Shapiro-Albert, In Prep]`, which is probably what led to some confusion.
