### 2020-10-21 66272a9
Added Py3.8 to Travis CI, and changed astropy version
