+++ setup.py
@@ -17,6 +17,7 @@ requirements = [
     'matplotlib>=2.0.0',
     'h5py',
     'pdat',
+    'astropy>=3.2,!=4.0.1,!=4.0.1.post1',
     'pint-pulsar',
 ]
 
@@ -58,9 +59,9 @@ setup(
         'License :: OSI Approved :: MIT License',
         'Natural Language :: English',
         'Programming Language :: Python :: 3',
-        'Programming Language :: Python :: 3.5',
         'Programming Language :: Python :: 3.6',
         'Programming Language :: Python :: 3.7',
+        'Programming Language :: Python :: 3.8',
     ],
     test_suite='tests',
     tests_require=test_requirements,
