### 2020-10-21 PsrSigSim/VersionZeroPointZero#156 @Hazboun6
I think we are going to need to just make a __strong__ note of the `Pint` namespace conflict. As @bshapiroalbert
noted, there is necessary functionality in the package and we are only going to require more of it as we add features.

There are ways around this, for instance we could change where `Pint` is [installed](https://stackoverflow.com/questions/27532112/how-to-handle-python-packages-with-conflicting-names), but my guess is actually that `pint-pulsar` is more likely to be installed by a user. One interesting point is that none of the developers use the more  popular `pint` because the `Quantity` functionality is largely duplicated in  `astropy`.

At least I can't think of a clean workaround.
