+++ requirements.txt
@@ -5,5 +5,4 @@ h5py>=2.7.0
 astropy>=2.0, <4.0
 pdat>=0.2
 fitsio==0.9.12
-packaging>=19.1
 pint-pulsar>=0.7
