### 2020-10-22 #2757 @mbobra
@Hazboun6 Good question. You don't need to cite the AAS paper. The DOI for the AAS paper will appear the JOSS paper automatically (see examples [here](https://joss.theoj.org/papers/10.21105/joss.01889) and [here](https://joss.theoj.org/papers/10.21105/joss.02125); look at the left margin).

There is some documentation about the AAS publishing process process [here](https://joss.readthedocs.io/en/latest/editing.html#handling-of-papers-published-together-with-aas-publishing). Here is the order of operations:

1. The JOSS reviewers agree that the software package PsrSigSim passes review.
2. Your paper is accepted by JOSS. However, we don't publish it yet. We pause the publication of the review pending the DOI from ApJ. Practically speaking, pausing the review means I add a label to this review thread that says "paused" and we leave this issue open.
3. Your paper is accepted into ApJ. We the publish the JOSS paper and automatically include the ApJ DOI, which automatically closes this issue.
