### 2020-10-23 ddffde7
Minor fixes due to deprication and other warnings
