### 2020-10-26 3e27830
use Nchan for both BB and FB
