### 2020-10-26 9f7ec1a
use correct signal types
