### 2020-10-26 639fc82
don't implement Telescope.observe() for BB signals
observing Baseband requires different telescope systems than FilterBank.  BB is raw voltage sampled at the ADC.  It's not yet passed to a pulsar backend
