### 2020-10-28 97cb98c
check signal type before `sublen`
