### 2020-10-28 a64bc7b
correct sigtype for BB and RF
