### 2020-11-01 d23fca1
JOSS BasebandSignal (#171)
* pass Telescope.Tsys to radiometer_noise()

* move nsamp to BaseSignal, populate for BasebandSignal too

* use Nchan for both BB and FB

* use correct signal types

* don't implement Telescope.observe() for BB signals

observing Baseband requires different telescope systems than FilterBank.  BB is raw voltage sampled at the ADC.  It's not yet passed to a pulsar backend

* test baseband

* check signal type before `sublen`

* correct sigtype for BB and RF

* point to correct methods in profile docstrings
