### 2020-11-02 36c2647
Added pytest warning filters for PINT Deprecation and FITSruntime warnings
