+++ tests/test_io.py
@@ -106,6 +106,7 @@ def test_savesig(PSRfits, pulsar):
 
 @pytest.mark.filterwarnings('ignore::fitsio.FITSRuntimeWarning')
 @pytest.mark.filterwarnings('ignore::DeprecationWarning')
+@pytest.mark.filterwarnings('ignore::astropy.utils.data.CacheMissingWarning')
 def test_savephaseconnect(PSRfits, pulsar):
     """
     Test getting a signal from a fits file, making pulses with it, and save it,
@@ -122,6 +123,7 @@ def test_savephaseconnect(PSRfits, pulsar):
 
 @pytest.mark.filterwarnings('ignore::fitsio.FITSRuntimeWarning')
 @pytest.mark.filterwarnings('ignore::DeprecationWarning')
+@pytest.mark.filterwarnings('ignore::astropy.utils.data.CacheMissingWarning')
 def test_savephaseconnect_inc(PSRfits, pulsar):
     """
     Test getting a signal from a fits file, making pulses with it, and save it,
@@ -165,6 +167,7 @@ def test_params(PSRfits, pulsar):
     assert(isinstance(PSRfits.stt_smjd.value, float))
 
 @pytest.mark.filterwarnings('ignore::DeprecationWarning')
+@pytest.mark.filterwarnings('ignore::astropy.utils.data.CacheMissingWarning')
 def test_gen_polyco(PSRfits):
     """
     Test generating the polyco dictionary.
@@ -185,6 +188,7 @@ def test_gen_metadata(pulsar, PSRfits):
     pri_dict, sub_dict = PSRfits._gen_metadata(S, pulsar, ref_MJD = 56000.0, inc_len = 10.0)
 
 @pytest.mark.filterwarnings('ignore::DeprecationWarning')
+@pytest.mark.filterwarnings('ignore::astropy.utils.data.CacheMissingWarning')
 def test_edit_header(pulsar):
     """
     Test editing the psrfits file header.
