### 2020-11-03 PsrSigSim/VersionZeroPointZero#164 @paulthebaker
Because `scipy.stats` uses `numpy.random` under the hood, a user can specify a random seed by calling `numpy.random.seed()` before running a function that uses a random number generator (RNG).    **We should add this fact to the documentation as part of the JOSS revisions.**

This is a legacy way to seed the RNG.  [There's a newer way to do this as of Numpy 1.17 using `BitGenerator` objects](https://numpy.org/doc/stable/reference/random/index.html), but the old way will still work for a while.  [One design motivation of the new way is to provide reproducible random samples across platforms for unit testing](https://numpy.org/neps/nep-0019-rng-policy.html).  For compatibility we might want to allow users to pass in integer seeds, older `RandomState` instances, or newer `BitGenerator` instances.

There are two places where we use a RNG, pulse generation (`pulsar._make_amp_pulses()` and `_make_pow_pulses()`) and noise generation (`reciever.radiometer_noise()`).  I could imagine someone might want to use identical pulses in new noise or keep the exact same noise but simulate new pulses.  In which case there should be at least three places to provide seeds: `pulsar.make_pulses()`, `telescope.observe()`, and a global seed (for an easy way to reseed a simulation).

Implementing seeding the "right" way is beyond the scope of the JOSS review, but we should do it eventually.  Getting the desired RNG infrastructure in place, will give an example for how contributors should implement RNG methods for other use cases.
