### 2020-11-13 PsrSigSim/PsrSigSim#175 @bshapiroalbert
Need user injected spectral index feature
We should add a feature to the `pulsar` class for users to put in a spectral index for the simulated pulsar if they want to, since we know that pulsars have rather steep spectral indices.
