### 2020-11-18 246c7bb
first round JOSS paper revisions
