+++ paper/paper.md
@@ -63,7 +63,7 @@ aas-journal: Astrophysical Journal
 
 Pulsar observations have been a hallmark of radio astronomy since Jocelyn Bell-Burnell
 discovered the repeating signal from PSR B1919+21 in 1967 [@hewish1968]. Today, most radio frequency observatories include dedicated equipment for observations of pulsars, e.g., @arecibo2019, @gbt2019, @chime2019, @meertime2020.
-The phenomenal precision of pulsar rotational periods has allowed for many notable observations stemming from strong field general relativity, including a Nobel prize for the discovery of the first binary neutron star [@taylor1975], which led to the first evidence for gravitational waves [@taylor1982].
+The phenomenal precision of pulsar rotational periods has allowed for many notable observations stemming from strong field general relativity, including a Nobel Prize for the discovery of the first binary neutron star [@taylor1975], which led to the first evidence for gravitational waves [@taylor1982].
 
 Pulsars continue to be used as probes of gravity as the constituents of pulsar timing arrays (PTAs). PTAs are collections of highly precise millisecond pulsars regularly
 monitored for shifts in their pulse arrival times indicative of gravitational
@@ -72,11 +72,13 @@ pulsar timing arrays and the astrophysics of nanohertz gravitational waves.
 
 # Statement of Need
 
-In order to use these neutron stars as a galactic-scale gravitational wave observatory it is important to understand the noise that affects the timing of radio pulses. The pulse times of arrival are modeled deterministically using aspects of the pulse profile, intrinsic flux, orbital parameters of the pulsar (if in a binary system), proper motion across the sky, effects of the interstellar medium (ISM), position of the observatory with respect to the solar system barycenter and various aspects of the telescope and backend processing.  Additionally, stochastic noise processes affect the times of arrival.  The end-to-end simulation of ``PsrSigSim`` provides a way to study the covariances of the timing model and noise terms in a controlled manner.
+In order to use neutron stars as a galactic-scale gravitational wave observatory, it is important to understand the noise that affects the timing of radio pulses. The pulse times of arrival are modeled deterministically using aspects of the pulse profile, intrinsic flux, orbital parameters of the pulsar (if in a binary system), proper motion across the sky, effects of the interstellar medium (ISM), position of the observatory with respect to the solar system barycenter and various aspects of the telescope and backend processing.  Additionally, stochastic noise processes affect the times of arrival.  The end-to-end simulation of ``PsrSigSim`` provides a way to study the covariances of the timing model and noise terms in a controlled manner.
 
-The ``PsrSigSim`` is a Python package for simulating radio telescope data from pulsar timing observations, largely based on the formalism presented in [Shapiro-Albert, 2020].
+Other packages? SigProc
+
+``PsrSigSim`` is a Python package for simulating radio telescope data from pulsar timing observations, largely based on the formalism presented in [@shapiro-albert+2020pss].
 It uses standard Python packages, such as ``Numpy`` [@numpy] and ``Astropy``
-[@astropy] to simulate radio pulses from a neutron star and the propagation effects upon that signal from the interstellar medium. ``Signal``-objects are passed to ``Pulsar``-objects and `Telescope`-objects to build a realistic pulsar signal from scratch. `ISM`-objects use various signal processing techniques, such as the Fourier shift theorem and convolution of pulse profiles with decaying exponentials, in order to simulate the frequency dependent effects of the interstellar medium on pulse arrival times. The ``PsrSigSim`` interfaces with various pulsar data standards, such as `PSRFITS`, to build data products that can be passed to existing data processing pipelines.
[…]
