+++ paper/paper.bib
@@ -34,6 +34,17 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@article{shapiro-albert+2020pss,
+	archiveprefix = {arXiv},
+	author = {Shapiro-Albert, B.~J. and {Hazboun}, J.~S. and {McLaughlin}, M.~A. and {Lam}, M.~T.},
+	eprint = {2010.07301},
+	journal = {arXiv},
+	keywords = {pta, noise mitigation, computational},
+	month = {oct},
+	primaryclass = {astro-ph.HE},
+	title = {A Study in Frequency-Dependent Effects on Precision Pulsar Timing Parameters with the Pulsar Signal Simulator},
+	year = {2020}}
+
 @article{numpy,
   author={S. {van der Walt} and S. C. {Colbert} and G. {Varoquaux}},
   doi={10.1109/MCSE.2011.37},
