### 2020-11-25 #2757 @Hazboun6
Hi @mbobra ! Thanks for checking in. We are making good progress. I think we're probably ~75% of the way through the issues. We also got the referee report from the AAS paper back and that also involved a few changes to the [code,](https://github.com/PsrSigSim/PsrSigSim/issues/175) so we've been working to add that functionality in as well.

I can update the referees with which Issues have been completed within a week, and I think we'll only have a few things left at that point.
