### 2020-12-01 c376364
Fixed bug where convolving a profile of zeros results in a profile of nans
