### 2020-12-01 165ee50
Fixed previous scattering bug fix
