+++ paper/paper.md
@@ -72,15 +72,15 @@ pulsar timing arrays and the astrophysics of nanohertz gravitational waves.
 
 # Statement of Need
 
-In order to use neutron stars as a galactic-scale gravitational wave observatory, it is important to understand the noise that affects the timing of radio pulses. The pulse times of arrival are modeled deterministically using aspects of the pulse profile, intrinsic flux, orbital parameters of the pulsar (if in a binary system), proper motion across the sky, effects of the interstellar medium (ISM), position of the observatory with respect to the solar system barycenter and various aspects of the telescope and backend processing.  Additionally, stochastic noise processes affect the times of arrival.  The end-to-end simulation of ``PsrSigSim`` provides a way to study the covariances of the timing model and noise terms in a controlled manner.
+In order to use neutron stars as a galactic-scale gravitational wave observatory, it is important to understand the noise that affects the timing of radio pulses. The pulse times of arrival are modeled deterministically using aspects of the pulse profile, intrinsic flux, orbital parameters of the pulsar (if in a binary system), proper motion across the sky, effects of the interstellar medium (ISM), position of the observatory with respect to the solar system barycenter and various aspects of the telescope and backend processing.  Additionally, stochastic noise processes affect the times of arrival.
 
-Other packages? SigProc
+A number of `c`/`c++`-based packages exist which have simulation capabilities. The `SigProc` package [@sigproc2011] has long been publicly available for processing radio-based pulsar observation data and has a simulation module able to use a timing model to reproduce data in its native `fb` data structure. ``PSRCHIVE`` [@hotan2004, @VanStraten2012], another popular software package for processing radio data of pulsars,  has the ability to produce simulated pulsar profiles, but not as part of a full simulation pipeline. There are timing packages [@Luo2020, @tempo2] that can simulate signals in the times of arrival of pulses, for instance to simulate gravitational wave signals, but these are a derivative data product making the injection of some noise processes and signals more difficult. The end-to-end simulation of ``PsrSigSim`` provides a way to study the covariances of the timing model and noise terms in a physically motivated and controlled manner.
 
-``PsrSigSim`` is a Python package for simulating radio telescope data from pulsar timing observations, largely based on the formalism presented in [@shapiro-albert+2020pss].
+``PsrSigSim`` is the first Python package for general purpose simulation of radio telescope data from pulsar timing observations, largely based on the formalism presented in [@shapiro-albert+2020pss].
 It uses standard Python packages, such as ``Numpy`` [@numpy] and ``Astropy``
-[@astropy] to simulate radio pulses from a neutron star and the effects of the interstellar material on the signal propagation. ``Signal`` objects are passed to ``Pulsar`` objects and `Telescope` objects to build a realistic pulsar signal from scratch. `ISM` objects use various signal-processing techniques, such as the Fourier shift theorem and convolution of pulse profiles with decaying exponentials, in order to simulate the frequency dependent effects of the interstellar medium on pulse arrival times. The ``PsrSigSim`` interfaces with various pulsar data standards, such as `PSRFITS`, to build data products that can be passed to existing data processing pipelines.
+[@astropy] to simulate radio pulses from a neutron star and the effects of the interstellar material on the signal propagation. ``Signal`` objects are passed to ``Pulsar`` objects and `Telescope` objects to build a realistic pulsar signal from scratch. `ISM` objects use various signal-processing techniques, such as the Fourier shift theorem and convolution of pulse profiles with decaying exponentials, in order to simulate the frequency-dependent effects of the interstellar medium on pulse arrival times. The ``PsrSigSim`` interfaces with various pulsar data standards, such as `PSRFITS` [@hotan2004], to build data products that can be passed to existing data-processing pipelines.
 
[…]
