+++ paper/paper.bib
@@ -280,3 +280,85 @@ archivePrefix = {arXiv},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2020PASA...37...28B},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
+
+@ARTICLE{hotan2004,
+	adsnote = {Provided by the SAO/NASA Astrophysics Data System},
+	adsurl = {https://ui.adsabs.harvard.edu/abs/2004PASA...21..302H},
+	archiveprefix = {arXiv},
+	author = {{Hotan}, A.~W. and {van Straten}, W. and {Manchester}, R.~N.},
+	doi = {10.1071/AS04022},
+	eprint = {astro-ph/0404549},
+	journal = {\pasa},
+	keywords = {pulsars: general, methods: data analysis, Astrophysics},
+	month = jan,
+	number = {3},
+	pages = {302-309},
+	primaryclass = {astro-ph},
+	title = {{PSRCHIVE and PSRFITS: An Open Approach to Radio Pulsar Data Storage and Analysis}},
+	volume = {21},
+	year = 2004,
+	Bdsk-Url-1 = {https://doi.org/10.1071/AS04022}
+}
+
+@MISC{sigproc2011,
+       author = {{Lorimer}, D.~R.},
+        title = "{SIGPROC: Pulsar Signal Processing Programs}",
+     keywords = {Software},
+         year = 2011,
+        month = jul,
+          eid = {ascl:1107.016},
+        pages = {ascl:1107.016},
+archivePrefix = {ascl},
+       eprint = {1107.016},
+          url = {http://sigproc.sourceforge.net/},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2011ascl.soft07016L},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+
+@ARTICLE{VanStraten2012,
+	adsnote = {Provided by the SAO/NASA Astrophysics Data System},
+	adsurl = {http://adsabs.harvard.edu/abs/2012AR%26T....9..237V},
+	archiveprefix = {arXiv},
+	author = {{van Straten}, W. and {Demorest}, P. and {Oslowski}, S.},
+	eprint = {1205.6276},
+	journal = {Astronomical Research and Technology},
+	keywords = {pulsar, data analysis, software},
+	month = jul,
+	pages = {237-256},
+	primaryclass = {astro-ph.IM},
+	title = {{Pulsar Data Analysis with PSRCHIVE}},
+	volume = 9,
+	year = 2012
+}
+
+
+@MISC{Luo2020,
+	adsnote = {Provided by the SAO/NASA Astrophysics Data System},
+	adsurl = {https://ui.adsabs.harvard.edu/abs/2019ascl.soft02007L},
+	archiveprefix = {ascl},
+	author = {{Luo}, Jing and {Ransom}, Scott and {Demorest}, Paul and {van Haasteren}, Rutger and {Ray}, Paul and {Stovall}, Kevin and {Bachetti}, Matteo and {Archibald}, Anne and {Kerr}, Matthew and {Colen}, Jonathan and {Jenet}, Fredrick},
+	eid = {ascl:1902.007},
+	eprint = {1902.007},
+	keywords = {Software},
+	month = feb,
+	pages = {ascl:1902.007},
+	title = {{PINT: High-precision pulsar timing analysis package}},
+	year = 2019
+}
+
+
+@misc{tempo2,
+	adsnote = {Provided by the SAO/NASA Astrophysics Data System},
+	adsurl = {https://ui.adsabs.harvard.edu/abs/2012ascl.soft10015H},
[…]
