+++ paper/paper.bib
@@ -34,7 +34,7 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@article{shapiro-albert+2020pss,
+@article{shapiroalbert2020pss,
 	archiveprefix = {arXiv},
 	author = {Shapiro-Albert, B.~J. and {Hazboun}, J.~S. and {McLaughlin}, M.~A. and {Lam}, M.~T.},
 	eprint = {2010.07301},
