+++ paper/paper.md
@@ -78,9 +78,11 @@ A number of `c`/`c++`-based packages exist which have pulsar signal simulation c
 
 ``PsrSigSim`` is the first Python package for general purpose simulation of radio telescope data from pulsar timing observations, based on the formalism presented in [@shapiroalbert2020pss].
 It uses standard Python packages, such as ``Numpy`` [@numpy] and ``Astropy``
-[@astropy] to simulate radio pulses from a neutron star and the effects of the interstellar material on the signal propagation. ``Signal`` objects are passed to ``Pulsar`` objects and `Telescope` objects to build a realistic pulsar signal from scratch. `ISM` objects use various signal-processing techniques, such as the Fourier shift theorem and convolution of pulse profiles with decaying exponentials, in order to simulate the frequency-dependent effects of the interstellar medium on pulse arrival times. The ``PsrSigSim`` interfaces with various pulsar data standards, such as `PSRFITS` [@hotan2004], to build data products that can be passed to existing data-processing pipelines.
+[@astropy] to simulate radio pulses from a neutron star and the effects of the interstellar material on the signal propagation. ``Signal`` objects are passed to ``Pulsar`` objects and `Telescope` objects to build a realistic pulsar signal from scratch. `ISM` objects use various signal-processing techniques, such as the Fourier shift theorem and convolution of pulse profiles with decaying exponentials, in order to simulate the frequency-dependent effects of the interstellar medium on pulse arrival times. ``PsrSigSim`` interfaces with various pulsar data standards, such as `PSRFITS` [@hotan2004], to build data products that can be passed to existing data-processing pipelines.
 
-The modularity of the simulation allows enumerable options for investigating the processing of pulsar signals using the built-in features. It also allows users to add new physical effects with ease. The signals from different pulsars can be treated by the same telescope in order to investigate how various pulsars would look using a particular telescope configuration. A user can either use a provided telescope/backend, or build one with their own specifications.  This modularity also makes the `PsrSigSim` an excellent tool for educating students about the physics of pulsar signals and how they are observed and processed.
+The modularity of the simulation allows enumerable options for investigating the processing of pulsar signals using the built-in features. It also allows users to add new physical effects with ease. The signals from different pulsars can be treated by the same telescope in order to investigate how various pulsars would look using a particular telescope configuration. A user can either use a provided telescope/backend, or build one with their own specifications.  This modularity also makes `PsrSigSim` an excellent tool for educating students about the physics of pulsar signals and how they are observed and processed.
+
+The generic functionality, modularity and the ability to add user-defined signals/noise make `PsrSigSim` an useful tool for all pulsar researchers.
 
 # Acknowledgements
 
