### 2020-12-18 PsrSigSim/VersionZeroPointZero#160 @Hazboun6
@JBorrow  I have made the suggested changes to the paper and I've now replaced the Shapiro-Albert reference with the arxiv version of the ApJ paper.
