### 2020-12-18 PsrSigSim/VersionZeroPointZero#161 @Hazboun6
@arjunsavel thank you again for these comments. I have addressed your comments and pushed the changes to the `joss-review` branch.
