### 2020-12-18 PsrSigSim/VersionZeroPointZero#164 @Hazboun6
@paulthebaker Am I interpreting this correctly to understand that, at least in regards to the JOSS review, this is in my court? I'm working on the docs, so this is easy for me to do.
