### 2020-12-18 78c945a
dial back numpy version
