+++ requirements.txt
@@ -1,4 +1,4 @@
-numpy>=1.16.4
+numpy==1.17.3
 scipy>=1.0.0
 matplotlib>=2.1.1
 h5py>=2.7.0
