### 2020-12-19 cfa1034
tweak install_requires
