+++ .travis.yml
@@ -7,6 +7,9 @@ python:
   - 3.7
   - 3.6
 
+env:
+  env: NUMPY=1.17
+
 # command to install dependencies, e.g. pip install -r requirements.txt --use-mirrors
 install:
   - pip install -U tox-travis
