+++ .travis.yml
@@ -15,5 +15,7 @@ install:
   - pip install -U tox-travis
   #- pip install git+https://github.com/nanograv/PINT.git@master
 
+before_script:
+- pip install -I numpy==1.17.3
 # command to run tests, e.g. python setup.py test
 script: tox
