### 2020-12-21 cedc810
Added test GitHub Actions CI
