### 2020-12-21 c248198
Delete ci_test.yml
