### 2020-12-21 c8ce8b3
Update ci_test.yml
Specify which branches to run tests on. Maybe this will force it to run?
