+++ .github/workflows/ci_test.yml
@@ -1,6 +1,11 @@
 name: Pulsar Signal Simulator
 
-on: [pull_request]
+on:
+  pull_request:
+    branches: 
+      - master
+      - joss-review
+      - psrfits_reform
 
 jobs:
   build:
