### 2020-12-21 7ce04da
Update ci_test.yml
That didn't work, removed specifying branches.
