### 2020-12-21 777baa1
Update ci_test.yml
Attempt to pass python 3.6 tests by not updating setuptools
