+++ .github/workflows/ci_test.yml
@@ -17,6 +17,11 @@ jobs:
       with:
         python-version: ${{ matrix.python-version }}
     - name: Install dependencies
+      uses: py-actions/py-dependency-install@v2
+      with:
+        update-pip: "true"
+        update-setuptools: "false"
+        update-wheel: "true"
       run: |
         python -m pip install --upgrade pip
         pip install flake8 pytest
