### 2020-12-21 50c36e8
Update ci_test.yml
That didn't work, trying to fix by reverting and updating setuptools as well as pip now.
