+++ .github/workflows/ci_test.yml
@@ -17,13 +17,9 @@ jobs:
       with:
         python-version: ${{ matrix.python-version }}
     - name: Install dependencies
-      uses: py-actions/py-dependency-install@v2
-      with:
-        update-pip: "true"
-        update-setuptools: "false"
-        update-wheel: "true"
       run: |
         python -m pip install --upgrade pip
+        python -m setuptools install --upgrade setuptools
         pip install flake8 pytest
         #if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
         pip install -r requirements.txt
