### 2020-12-21 460bd65
Update ci_test.yml
Changed update to setuptools to hopefully run..
