+++ .github/workflows/ci_test.yml
@@ -18,8 +18,7 @@ jobs:
         python-version: ${{ matrix.python-version }}
     - name: Install dependencies
       run: |
-        python -m pip install --upgrade pip
-        python -m setuptools install --upgrade setuptools
+        python -m pip install --upgrade pip setuptools wheel
         pip install flake8 pytest
         #if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
         pip install -r requirements.txt
