+++ README.rst
@@ -6,6 +6,7 @@ PsrSigSim
 .. image:: https://img.shields.io/pypi/v/psrsigsim.svg
         :target: https://pypi.python.org/pypi/psrsigsim
 
+[![Actions Status](https://github.com/PsrSigSim/PsrSigSim/workflows/PsrSigSim%20Build/badge.svg)](https://github.com/PsrSigSim/PsrSigSim/actions)
 .. image:: https://img.shields.io/travis/PsrSigSim/PsrSigSim.svg
         :target: https://travis-ci.org/PsrSigSim/PsrSigSim
 
