### 2020-12-22 PsrSigSim/VersionZeroPointZero#175 @bshapiroalbert
This has been merged into the `psrfits_reform` and `joss-review` branches, but still needs to eventually be merged into the main branch.
