+++ HISTORY.rst
@@ -4,7 +4,7 @@ History
 1.0.0 (2020-10-15)
 ------------------
 
-*First release, coinciding with chromatic timing parameter paper.
+* First release, coinciding with chromatic timing parameter paper.
 
 0.2.0 (2020-08-17)
 ------------------
@@ -12,6 +12,6 @@ History
 * Pre Release version for testing.
 
 0.1.0 (2018-01-16)*
-------------------
+-------------------
 
 * First release on PyPI.
