### 2021-01-08 PsrSigSim/VersionZeroPointZero#158 @Hazboun6
Closing after consulting with @bshapiroalbert
