### 2021-01-08 5e50f18
docs edits for JOSS review
