+++ docs/pulse_profiles_tutorial_3.rst
@@ -1,3 +1,5 @@
+
+
 .. note:: This tutorial was generated from a Jupyter notebook that can be
           downloaded `here <_static/notebooks/pulse_profiles_tutorial_3.ipynb>`_.
 
@@ -55,7 +57,7 @@ data release.
     sublen = 60.0 # subintegration length in seconds, or rate to dump data at
     # Now we define our signal
     signal_1713 = pss.signal.FilterBankSignal(fcent = f0, bandwidth = bw, Nsubband=Nf, sample_rate = f_samp,
-                                           sublen = sublen, fold = True) # fold is set to `True`
+                                              sublen = sublen, fold = True) # fold is set to `True`
 
 
 .. parsed-literal::
@@ -86,7 +88,7 @@ Pulse Profiles
 In previous tutorials, we have defined a very simple Gaussian profile as
 the pulse profile. However, the ``PsrSigSim`` allows users to define
 profiles in a few different ways, including multiple Gaussians, a user
-input profile in the form of a Python array, and two dimensional
+input profile in the form of a Python array, and two-dimensional
 versions of the pulse profiles called pulse portaits.
 
 We will go through a few different ways to set up the pulse profiles,
@@ -100,7 +102,7 @@ The first method is the Gaussian profile, which has been demonstrated in
 previous tutorials. The Guassian needs three parameters, an amplitude, a
 width (or sigma), and a peak, the center of the Gaussian in phase space
 (e.g. 0-1). The simplest profile that can be defined is a single
-Gauassian.
+Gaussian.
 
 .. code:: python
 
@@ -120,6 +122,7 @@ and then we can give it a number of phase bins and plot it.
     # And then we can plot the array to see what the profile looks like
     plt.plot(np.linspace(0,1,2048), gauss_prof.profiles[0])
     plt.xlabel("Phase")
+    plt.ylabel('Normalized Intensity')
     plt.show()
     plt.close()
 
@@ -127,6 +130,19 @@ and then we can give it a number of phase bins and plot it.
 
 .. image:: pulse_profiles_tutorial_3_files/pulse_profiles_tutorial_3_11_0.png
 
+
+\* A Note on Intensity
+~~~~~~~~~~~~~~~~~~~~~~
+
+Pulsar astronomers often use arbitrary units for radio flux and
+intensity (~flux^2) when looking at pulsar timing data. For purposes of
+pulsar timing, while the actual signal-to-noise ratio obviously has an
+effect on the precision of times-of-arrival measurements, in diagnostic
+plots the data is often not converted to actual intensity. In many of
+these tutorials we follow this practice, and will specify if it is
+different. For more details about flux/intensity in pulsar data see the
+`Hand Book of Pulsar
+Astronomy <https://books.google.com/books/about/Handbook_of_Pulsar_Astronomy.html?id=OZ8tdN6qJcsC>`__.
 
 However the Gaussian profile can also be used to make a pulse profile
 with multiple Gaussian components. Instead of inputting a single value
@@ -147,18 +163,19 @@ array, it cannot be a list.
     amps = np.array([1.0, 0.1, 0.5])
     
     # Define the profile using multiple Gaussians
-    mulit_gauss_prof = pss.pulsar.GaussProfile(peak = peaks, width = widths, amp = amps)
+    multi_gauss_prof = pss.pulsar.GaussProfile(peak = peaks, width = widths, amp = amps)
[…]
