### 2021-01-09 PsrSigSim/VersionZeroPointZero#162 @Hazboun6
@arjunsavel At this point I have made all of the suggested changes and corrections above. Changes have been pushed to the `joss-review` branch and will be in the main branch shortly.
