### 2021-01-11 #2757 @Hazboun6
@mbobra I keep holding back on answering you thinking that I will be done with all of the changes we'd like to make. We are _almost_ done on our end. I spent a good amount of time on the docs today and I have been closing the various Issues in our GitHub repo. I don't think we need any help at this point but I appreciate the offer. Is there a hard deadline for these responses?
