### 2021-01-11 #2757 @mbobra
@Hazboun6 Nope, no hard deadline. I wanted to check in to make sure everything is okay. Please take all the time you need.

If you need a lot of time, that is totally fine -- but I in that case, I would pause the submission. All this does is notify the Editors in Chief that this submission will take a bit longer. Practically, all I do is put another label on this issue that says "Paused." In any case -- if we're looking at 4 weeks or less, no worries. If you think you'll need more time, I'll add the paused button (which we can unpause at any time).
