### 2021-01-27 PsrSigSim/VersionZeroPointZero#159 @Hazboun6
@JBorrow I've addressed all of the comments above, making changes in the relevant parts of the documentation. The only piece I have not addressed is adding in more use of Sphinx domains. I've opened an issue #185 . It seems like it would be better done as part of a larger restructuring of the documentation.
