### 2021-01-27 652b9ae
added not on random seed to tutorials
