### 2021-01-27 #2757 @Hazboun6
@mbobra  we have completed the changes to the code suggested by @JBorrow and @arjunsavel.

We branched the code and just merged all of the changes in this [PR](https://github.com/PsrSigSim/PsrSigSim/pull/187).

Thank you for your patience in how long this took us to do!
