### 2021-01-28 #2757 @Hazboun6
@JBorrow please standby. I just realized that the docs did not build on `readthedocs`. The were of course building locally. How embarrassing... My work day starts in an hour and I'll fix ASAP. Sorry
