### 2021-01-28 754efa4
Update requirements.txt
Trying this simple fix for `readthedocs`
