+++ docs/requirements.txt
@@ -1,4 +1,4 @@
-numpy>=1.16.4
+numpy>=1.16.4, <1.20
 scipy>=1.0.0
 matplotlib>=2.1.1
 astropy>=2.0
