### 2021-02-01 PsrSigSim/PsrSigSim#189 @arjunsavel
Further documentation edits
Thank you for making the previously requested changes to the documentation — things look great! I especially appreciate the random number notice at the bottom of the notebooks. A few further items to address:

- [x] Infrequent usage of psrsimsig vs. psrsigsim, found in
        - index.rst
        - constributing.rst (the GitHub link)
        - installation.rst
- [x] In contributing, the Travis link should now point to https://github.com/PsrSigSim/PsrSigSim/actions, I believe
- [x] in the random generation note:
        - a call the seed generator —> a call to the seed generator
        - is essentially used —> are essentially used

This issue is part of my JOSS review.
