### 2021-02-01 PsrSigSim/PsrSigSim#190 @arjunsavel
Further paper edits
The JOSS paper is shaping up nicely! Just a couple more minor comments:

- [x] The multiple parenthetical citations ([@hotan2004, @VanStraten2012], [@Luo2020, @tempo2]) aren't being rendered correctly; I believe the individual citations should be separated by semicolons instead of commas.
- [x] Bell-Burnell  —> Bell Burnell
- [x] Numpy —> NumPy
- [x] Comma after Astropy citation
- [x] general purpose —> general-purpose

This issue is part of my JOSS review.
