### 2021-02-05 #2757 @Hazboun6
@arjunsavel Thanks for the second look at the paper and docs! I've merged changes based on your suggestions and regenerated the pdf above. Also looks like the semicolon fixed the multiple citations.
