### 2021-02-08 #2757 @mbobra
Thank you both so very much for your thorough reviews, @arjunsavel and @JBorrow! Your volunteer time keeps JOSS running and the JOSS team appreciates it very much ☀️
