### 2021-02-08 #2757 @mbobra
@Hazboun6 Could you please cut a new release of PsrSigSim and [deposit it in Zenodo](https://guides.github.com/activities/citable-code/)? Please make sure that the author list on the Zenodo deposit matches the author list on the paper. Then I can move forward with accepting the paper.
