### 2021-02-08 #2757 @Hazboun6
@mbobra I just tagged a release and deposited in [Zenodo](https://zenodo.org/record/4521235). Should be all set. Thanks to everyone for helping make our code better!
