### 2021-02-09 #2757 @Hazboun6
@mbobra do you need any info about the ApJ paper? It has been excepted and the proofs were returned yesterday, but hasn't been published yet.
