### 2021-02-09 #2757 @mbobra
@Hazboun6 Yes! I was just about to message you with that. Please ask AAS to give you the DOI for the paper (I think they should have already generated that) and then put it in the paper.md YAML header [like this](https://joss.readthedocs.io/en/latest/editing.html#handling-of-papers-published-together-with-aas-publishing). Please let me know when that's done and I can wrap up here.
