### 2021-02-12 #2757 @mbobra
@Hazboun6 Congrats!!! All we have to do now is wait for the ApJ paper to be published. Then we can publish this one (which takes just a few minutes). Can you keep us up to date about when the ApJ paper is published?

@openjournals/joss-eics This paper, which is a AAS companion paper, is ready to publish.
