### 2021-02-13 #2757 @Hazboun6
@arfon Yep! It is "A Study in Frequency-dependent Effects on Precision Pulsar Timing Parameters with the Pulsar Signal Simulator"
