### 2021-02-14 #2757 @Hazboun6
The DOI is `10.3847/1538-4357/abdc29` The alphanumeric past the last slash is the AAS submission ID as well. And I've put the DOI into the `paper.md` in the appropriate spot.
