### 2021-02-14 #2757 @arfon
> @Hazboun6 Congrats!!! All we have to do now is wait for the ApJ paper to be published. Then we can publish this one (which takes just a few minutes). Can you keep us up to date about when the ApJ paper is published?

@mbobra - as I think we have the AAS DOI now, we can go ahead and accept and publish here too.
